expert_search/Avito/requirements.txt
a2wsgi==1.10.8
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.2
apache-airflow-core==3.0.2
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.2
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cadwyn==5.3.3
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.11
DAWG2-Python==0.9.0
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.1.1
distro==1.9.0
dnspython==2.7.0
dynaconf==3.2.11
email_validator==2.2.0
emoji==2.14.1
evidently==0.7.7
Faker==37.4.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
Flask==2.2.5
fonttools==4.58.4
fsspec==2025.5.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.4.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.11.0
libcst==1.8.0
linkify-it-py==2.0.3
litestar==2.16.0
litestar-htmx==0.5.0
lockfile==0.12.2
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
multipart==1.2.1
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
patsy==1.0.1
pendulum==3.1.0
pillow==11.2.1
plotly==5.24.1
pluggy==1.6.0
polyfactory==2.21.0
preshed==3.0.10
protobuf==4.25.8
psutil==7.0.0
psycopg2==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.2.3
PySocks==1.7.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retryhttp==1.3.3
rich==13.9.4
rich-argparse==1.7.1
rich-click==1.8.9
rich-toolkit==0.14.7
rpds-py==0.25.1
ru_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.8.0/ru_core_news_lg-3.8.0-py3-none-any.whl#sha256=90bd584be86772b647e1b15cdc62e09d3a7182284ee23da95e19c2bf71aa0189
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
sentence-transformers==4.1.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
srsly==2.5.1
starlette==0.46.2
statsmodels==0.14.4
structlog==25.4.0
svcs==25.1.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
types-requests==2.32.0.20250602
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
ujson==5.10.0
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.3
uvloop==0.21.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.2
zipp==3.22.0
